fix(user): cannot access to methods order than register

+++ apps/upload/src/comsumer.upload.ts
@@ -22,45 +22,45 @@ export class ConsumerService implements OnModuleInit {
           const messageData = JSON.parse(message.content.toString());
           const cmd = messageData.cmd;
           const payload = messageData.payload;
-          console.log(cmd, payload);
+          if (cmd != undefined && payload != undefined) {
+            //validate file if have space change to -
 
-          switch (cmd) {
-            case UploadMethod.UploadSingle:
-              await this.uploadService.upload(payload.fileName, payload.file);
-              break;
-            case UploadMethod.UploadMultiple:
-              await this.uploadService.uploadMultiple(payload);
-              break;
-            case UploadMethod.Update:
-              await this.uploadService.update(
-                payload.fileName,
-                payload.file,
-                payload.oldFileName,
-              );
-              break;
-            case UploadMethod.Delete:
-              await this.uploadService.delete(payload.fileName);
-              break;
-            case UploadMethod.DeleteMultiple:
-              await this.uploadService.deleteMultiple(payload);
-              break;
-            default:
-              this.logger.error('Unknown command:', cmd);
-              break;
+            if (payload.fileName.includes(' ')) {
+              const newFileName = payload.fileName.replaceAll(' ', '-');
+              payload.fileName = newFileName;
+            }
+            const {
+              fileName,
+              file: { data },
+            } = payload;
+            const dataUpload = Buffer.from(data);
+            switch (cmd) {
+              case UploadMethod.UploadSingle:
+                await this.uploadService.upload(fileName, dataUpload);
+                break;
+              case UploadMethod.UploadMultiple:
+                await this.uploadService.uploadMultiple(payload);
+                break;
+              case UploadMethod.Update:
+                await this.uploadService.update(
+                  fileName,
+                  dataUpload,
+                  payload.oldFileName,
+                );
+                break;
+              case UploadMethod.Delete:
+                await this.uploadService.delete(fileName);
+                break;
+              case UploadMethod.DeleteMultiple:
+                await this.uploadService.deleteMultiple(payload);
+                break;
+              default:
+                this.logger.error('Unknown command');
+                break;
+            }
           }
-          channel.ack(message);
 
-          // switch (message?.content.toString()) {
-          //   case :
-          // }
-          // if (message) {
-          //   // console.log(message);
-          //   console.log(message.content.toString());
-          //   const content = JSON.parse(message.content.toString());
-          //   this.logger.log('Received message:', content);
-          //   // await this.emailService.sendEmail(content);
-          //   channel.ack(message);
-          // }
+          channel.ack(message);
         });
       });
       this.logger.log('Consumer service started and listening for messages.');
